Refactor: Remove cal_nelec and cal_nbands from ucell.cpp.
